% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdLISA.R
\name{pdLISA}
\alias{pdLISA}
\title{Product density LISA functions}
\usage{
pdLISA(xy, s.region, ds, ks = "box", hs)
}
\arguments{
\item{xy}{Spatial coordinates \eqn{(x,y)} of the point pattern.}

\item{s.region}{A two-column matrix specifying a polygonal region containing all data locations. If \code{s.region} is missing, the Ripley-Rasson estimate convex spatial domain is considered.}

\item{ds}{A vector of distances \code{u} at which \eqn{\rho^{(2)i}(u)} is computed.}

\item{ks}{A kernel function for the spatial distances. The default is the \code{"box"} kernel. It can also be \code{"epanech"} for the Epanechnikov kernel, or \code{"biweight"}.}

\item{hs}{A bandwidth of the kernel function \code{ks}.}
}
\value{
A list containing:
\itemize{
  \item \code{lisa}: A vector containing the values of \eqn{\widehat{\rho}^{(2)i}(r)} estimated.
  \item \code{ds}: If \code{ds} is missing, a vector of distances \code{u} at which \eqn{\rho^{(2)i}(u)} is computed under the restriction \eqn{0<\epsilon<r}.
  \item \code{kernel}: A vector of names and bandwidth of the spatial kernel.
  \item \code{s.region}: Parameter passed in argument.
  }
}
\description{
Computes an edge-corrected kernel estimator of the product density LISA functions.
}
\examples{
## Not run:
#################

# Realisations of the homogeneous Poisson processes
pp <- rpoispp(100)
xy <- cbind(pp$x,pp$y)

# R plot
plot(xy,xlab="x",ylab="y")

# This function provides an edge-corrected kernel estimator of the porduct density LISA functions.
out <- pdLISA(xy)
out

# R plot - Temporal mark variogram
par(mfrow=c(1,1))
plot(out$ds,out$lisa[1,], type="l", xlab="dist", ylab="LISA",
ylim=(c(min(out$lisa),max(out$lisa))),main="Product density LISA functions")
for (i in 2:length(xy[,1])){lines(out$ds,out$lisa[i,])}
lines(out$ds,rep((length(xy[,1])-2)*(length(xy[,1])-1)+length(xy[,1]),length(out$ds)),col="red")

## End(Not run)
}
\author{
Francisco J. Rodriguez-Cortes <cortesf@uji.es> \url{https://fjrodriguezcortes.wordpress.com}
}
\references{
Baddeley, A. and Turner, J. (2005). \code{spatstat}: An R Package for Analyzing Spatial Point Pattens. Journal of Statistical Software 12, 1-42.

Cressie, N. and Collins, L. B. (2001). Analysis of spatial point patterns using bundles of product density LISA functions. Journal of Agricultural, Biological, and Environmental Statistics 6, 118-135.

Cressie, N. and Collins, L. B. (2001). Patterns in spatial point locations: Local indicators of spatial association in a minefield with clutter Naval Research Logistics (NRL), John Wiley & Sons, Inc. 48, 333-347.

Stoyan, D. and Stoyan, H. (1994). Fractals, random shapes, and point fields: methods of geometrical statistics. Chichester: Wiley.
}

