% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.numbers.R
\name{plot.numbers}
\alias{plot.numbers}
\title{Plot of label points}
\usage{
plot.numbers(pattern, ...)
}
\arguments{
\item{pattern}{The spatial point pattern to be plotted. An object of class "ppp".}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par). Many methods will accept the following arguments: type
what type of plot should be drawn. Possible types are
\code{"p"} for points,
\code{"l"} for lines,
\code{"b"} for both,
\code{"c"} for the lines part alone of "b",
\code{"o"} for both ‘overplotted’,
\code{"h"} for ‘histogram’ like (or ‘high-density’) vertical lines,
\code{"s"} for stair steps,
\code{"S"} for other steps, see ‘Details’ below,
\code{"n"} for no plotting.
All other types give a warning or an error; using, e.g., type = "punkte" being equivalent to type = "p" for S compatibility. 
Note that some methods, e.g. plot.factor, do not accept this. 
\code{main}
an overall title for the plot: see title.
\code{sub}
a sub title for the plot: see title.
\code{xlab}
a title for the x axis: see title.
\code{ylab}
a title for the y axis: see title.
\code{asp}
the y/x aspect ratio, see plot.window.}
}
\value{
(Invisible) object of class "symbolmap" giving the correspondence between labeled points and plotting characters.
}
\description{
Draws a text label at a given x and y coordinate into a spatial point pattern.
}
\examples{
## Not run:
#################

# Realisations of the homogeneous Poisson processes
pattern <- humberside
plot.numbers(pattern, main = "Child Leukaemia")

## End(Not run)
}
\author{
Jonatan A Gonzalez
}

