% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdLISA.R
\name{pdLISA}
\alias{pdLISA}
\title{Product density LISA functions}
\usage{
pdLISA(xy, s.region, ds, ks = "epanech", hs)
}
\arguments{
\item{xy}{Spatial coordinates \eqn{(x,y)} of the point pattern.}

\item{s.region}{A two-column matrix specifying a polygonal region containing all data locations. If \code{s.region} is missing, the Ripley-Rasson estimate convex spatial domain is considered.}

\item{ds}{A vector of distances \code{u} at which \eqn{\rho^{(2)i}(u)} is computed.}

\item{ks}{A kernel function for the spatial distances. The default is \code{"epanech"} the Epanechnikov kernel. It can also be \code{"box"} kernel, or \code{"biweight"}.}

\item{hs}{A bandwidth of the kernel function \code{ks}.}
}
\value{
A list containing:
\itemize{
  \item \code{lisa}: A matrix containing the values of the estimation of \eqn{\widehat{\rho}^{(2)i}(r)}  for each one of \eqn{n} points of the process by rows.
  \item \code{ds}: If \code{ds} is missing, a vector of distances \code{u} at which \eqn{\rho^{(2)i}(u)} is computed under the restriction \eqn{0<\epsilon<r}.
  \item \code{kernel}: A vector of names and bandwidth of the spatial kernel.
  \item \code{s.region}: Parameter passed in argument.
  }
}
\description{
Computes an edge-corrected kernel estimator of the product density LISA functions.
}
\details{
An individual product density LISA functions \eqn{\rho^{(2)i}(.)} should reveal the extent of the contribution of the event \eqn{u_i} to the global estimator of the second-order product density \eqn{\rho^{(2)}(.)}, and may provide a further description of structure in the data (e.g., determining events with similar local structure through dissimilarity measures of the individual LISA functions), for more details see Cressie and Collins (2001).
}
\examples{
## Not run:
#################

# Realisations of the homogeneous Poisson processes
pp <- rpoispp(100)
plot(pp)

xy <- cbind(pp$x,pp$y)

# This function provides an edge-corrected kernel estimator of the porduct density LISA functions.
out <- pdLISA(xy)
out

# R plot
par(mfrow=c(1,1))
plot(out$ds,out$lisa[1,], type="l", xlab="dist", ylab="LISA",
ylim=(c(min(out$lisa),max(out$lisa))),main="Product density LISA functions")
for (i in 2:length(xy[,1])){lines(out$ds,out$lisa[i,])}
lines(out$ds,rep((length(xy[,1])-2)*(length(xy[,1])-1)+(length(xy[,1])-1),length(out$ds)),col="red")

## End(Not run)
}
\author{
Francisco J. Rodriguez-Cortes <cortesf@uji.es> \url{https://fjrodriguezcortes.wordpress.com}
}
\references{
Baddeley, A. and Turner, J. (2005). \code{spatstat}: An R Package for Analyzing Spatial Point Pattens. Journal of Statistical Software 12, 1-42.

Cressie, N. and Collins, L. B. (2001). Analysis of spatial point patterns using bundles of product density LISA functions. Journal of Agricultural, Biological, and Environmental Statistics 6, 118-135.

Cressie, N. and Collins, L. B. (2001). Patterns in spatial point locations: Local indicators of spatial association in a minefield with clutter Naval Research Logistics (NRL), John Wiley & Sons, Inc. 48, 333-347.

Stoyan, D. and Stoyan, H. (1994). Fractals, random shapes, and point fields: methods of geometrical statistics. Chichester: Wiley.
}

